requirements.txt
# ─────────────────────────────────────────────
# Requirements — Analyseur Contrats Assurance-Crédit
# ─────────────────────────────────────────────

# Interface
streamlit>=1.35.0

# Extraction PDF
pdfplumber>=0.11.0
PyMuPDF>=1.24.0          # fitz

# DOCX
python-docx>=1.1.0

# OCR (optionnel — nécessite Tesseract installé sur le système)
# pytesseract>=0.3.10
# pdf2image>=1.17.0

# NLP / IA
anthropic>=0.30.0

# Données
pandas>=2.2.0

# Excel
openpyxl>=3.1.0
